Make use of utils consistent
