+++ bokeh/charts/builder/tests/test_bar_builder.py
@@ -22,7 +22,8 @@ import pandas as pd
 
 from bokeh.charts import Bar
 from bokeh.models import Range1d, FactorRange
-from _utils import create_chart
+from bokeh.util.testing import create_chart
+
 
 #-----------------------------------------------------------------------------
 # Classes and functions
