+++ bokeh/charts/builder/tests/test_step_builder.py
@@ -22,8 +22,7 @@ from numpy.testing import assert_array_equal
 import pandas as pd
 
 from bokeh.charts import Step
-
-from _utils import create_chart
+from bokeh.util.testing import create_chart
 
 #-----------------------------------------------------------------------------
 # Classes and functions
