+++ bokeh/charts/builder/tests/test_timeseries_builder.py
@@ -23,8 +23,8 @@ from numpy.testing import assert_array_equal
 import pandas as pd
 
 from bokeh.charts import TimeSeries
+from bokeh.util.testing import create_chart
 
-from _utils import create_chart
 
 #-----------------------------------------------------------------------------
 # Classes and functions
