+++ bokeh/util/testing.py
@@ -159,3 +159,32 @@ class FlaskClientTestCase(BaseBokehServerTestCase):
 
     def tearDown(self):
         pass
+
+#----------------------
+# For testing charts
+#----------------------
+
+def create_chart(klass, values, compute_values=True, **kws):
+    """ Create a new chart klass instance with values and the extra kws keyword
+    parameters.
+
+    Args:
+        klass (class): chart class to be created
+        values (iterable): chart data series
+        compute_values (bool): if == True underlying chart attributes (like data,
+                ranges, source, etc..) are computed by calling _setup_show,
+                _prepare_show and _show_teardown methods.
+        **kws (refer to klass arguments specification details)
+
+    Return:
+        _chart: klass chart instance
+    """
+    _chart = klass(
+        values, title="title", xlabel="xlabel", ylabel="ylabel",
+        legend="top_left", xscale="linear", yscale="linear",
+        width=800, height=600, tools=True,
+        filename=False, server=False, notebook=False,
+        **kws
+    )
+
+    return _chart
